requirements.txt
# validation_simulation.py 実行に必要なライブラリ

# データ処理・分析
numpy>=1.21.0
pandas>=1.3.0

# 地理空間インデックス
h3>=3.7.0

# プログレスバー
tqdm>=4.62.0

# データ可視化
matplotlib>=3.5.0
seaborn>=0.11.0

# YAML設定ファイル
PyYAML>=6.0

# 機械学習・深層学習
torch>=1.9.0
torchvision>=0.10.0

# 日本語フォント対応（Windows環境用）
# Windowsの場合、matplotlibで日本語フォント（Meiryo）を使用するため
# 必要に応じて以下もインストール:
# japanize-matplotlib>=1.1.3 